+++ packages/tinacms/src/toolkit/fields/components/reference/model/reference-field-props.ts
@@ -1,4 +1,4 @@
-import { CollectionFilters } from '../utils/fetch-optioins-query-builder'
+import { CollectionFilters } from '../utils/fetch-options-query-builder'
 
 export type Option = {
   value: string
